Changes to adopt new write and read request/response formats.
When writing the sender makes a POST request with content-type
application/octet-stream and when reading the response (when data is read
successfully) has the same content-type instead of a json-encoded message where
the "data" has to be encoded again in base64 (or whatever).

Close method also updated to use new DELETE http method instead of POST with
special body.